Reduce ChatRoomTTL to uint32 for easy int64 conversion.

+++ server/cmd/api/chatconfig.go
@@ -1,7 +1,7 @@
 package main
 
 type ChatConfig struct {
-	ChatRoomTTL                 uint64 `json:"chatRoomTTL"`
+	ChatRoomTTL                 uint32 `json:"chatRoomTTL"`
 	MaxNumberOfParticipants     uint64 `json:"maxNumberOfParticipants"`
 	MaxHistoryDurationInSeconds uint64 `json:"maxHistoryDurationInSeconds"`
 	MaxMessagesLength           uint64 `json:"maxMessagesLength"`
